Renamed Search to SearchPage
