+++ src/pages/Entity/EntityPage.tsx
@@ -15,7 +15,7 @@ import withAccessControl from '../../services/withAccessControl';
 import Axios from 'axios';
 import EdsIcon from '../../components/icons/EdsIcon';
 import { COLORS } from '../../style/GlobalStyles';
-import { SearchType } from '../Search/Search';
+import { SearchType } from '../Search/SearchPage';
 import WorkOrderInfo from './WorkOrderInfo';
 import {
     BackButton,
