+++ src/services/apiTypeGuards.ts
@@ -1,6 +1,6 @@
 import { CheckItem } from '@equinor/procosys-webapp-components/dist/typings/apiTypes';
 import { SavedSearchType } from '../pages/Search/SavedSearches/SavedSearchResult';
-import { SearchType } from '../pages/Search/Search';
+import { SearchType } from '../pages/Search/SearchPage';
 import {
     ChecklistDetails,
     ChecklistResponse,
